Profile: Reorganize Vet360 containers and child components into vet360 directory (#8066)
* Add the base vet360 form components

* Correct the prop types

* Move over the reducer

* Update reducers for portability; convert address actions to constants

* Remove all references to address constants

* Derive the analytics section rather than passing as a prop

* Remove missed references

* Update tests

* Move Vet360 field/sections into vet360 directory and update references

* Move field views into separate files

* Move over the selectors

* Remove now-redundant "Vet360" prefix on file names

* Updates tests

* Add back the generated prop analyticsSectionName for child components
