--- src/applications/personalization/profile360/components/AddressSection.jsx
@@ -1,47 +1,0 @@
-import React from 'react';
-
-import {
-  isEmptyAddress,
-  formatAddress
-} from '../../../../platform/forms/address/helpers';
-
-import Vet360ProfileField from '../containers/Vet360ProfileField';
-import AddressEditModal from './AddressEditModal';
-
-function AddressView({ data: address }) {
-
-
-  const { street, cityStateZip, country } = formatAddress({
-    addressOne: address.addressLine1,
-    addressTwo: address.addressLine2,
-    addressThree: address.addressLine3,
-    // force formatting of military addresses as domestic
-    type: address.addressType.match(/military/i) ? 'DOMESTIC' : address.addressType.toUpperCase(),
-    ...address,
-  });
-
-  return (
-    <div>
-      {street}<br/>
-      {cityStateZip}<br/>
-      {country}
-    </div>
-  );
-}
-
-function isEmpty({ data: addressData } = {}) {
-  return isEmptyAddress(addressData);
-}
-
-export default function Vet360Address({ title, fieldName, analyticsSectionName, deleteDisabled }) {
-  return (
-    <Vet360ProfileField
-      title={title}
-      fieldName={fieldName}
-      analyticsSectionName={analyticsSectionName}
-      deleteDisabled={deleteDisabled}
-      isEmpty={isEmpty}
-      Content={AddressView}
-      EditModal={AddressEditModal}/>
-  );
-}
