+++ src/applications/personalization/profile360/vet360/components/EmailField/View.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function EmailView({ data: emailData }) {
+  return <span>{emailData.emailAddress}</span>;
+}
