+++ src/applications/personalization/profile360/vet360/components/EmailField/index.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-import Vet360ProfileField from '../containers/Vet360ProfileField';
-import EmailEditModal from './EmailEditModal';
+import Vet360ProfileField from '../../containers/ProfileField';
+import EmailEditModal from './EditModal';
 
-function EmailView({ data: emailData }) {
-  return <span>{emailData.emailAddress}</span>;
-}
+import EmailView from './View';
 
 export default function Vet360Email({ title = 'Email address', fieldName = 'email', analyticsSectionName = 'email' }) {
   return (
