+++ src/applications/personalization/profile360/vet360/components/base/Transaction.jsx
@@ -5,10 +5,10 @@ import classNames from 'classnames';
 import {
   isPendingTransaction,
   isFailedTransaction
-} from '../util/transactions';
+} from '../../../util/transactions';
 
-import Vet360TransactionInlineErrorMessage from './Vet360TransactionInlineErrorMessage';
-import Vet360TransactionPending from './Vet360TransactionPending';
+import Vet360TransactionInlineErrorMessage from './TransactionInlineErrorMessage';
+import Vet360TransactionPending from './TransactionPending';
 
 export default class Vet360Transaction extends React.Component {
   static propTypes = {
