+++ src/applications/personalization/profile360/vet360/components/base/TransactionErrorBanner.jsx
@@ -5,7 +5,7 @@ import {
   hasGenericUpdateError,
   hasMVIError,
   hasMVINotFoundError
-} from '../util/transactions';
+} from '../../../util/transactions';
 
 export function GenericUpdateError(props) {
   return (
