+++ src/applications/personalization/profile360/vet360/containers/PendingTransactionCategory.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { connect } from 'react-redux';
 import AlertBox from '@department-of-veterans-affairs/formation/AlertBox';
 
-import { refreshTransaction } from '../actions/updaters';
+import { refreshTransaction } from '../../actions/updaters';
 
-import Vet360TransactionPending from '../components/Vet360TransactionPending';
+import Vet360TransactionPending from '../components/base/TransactionPending';
 
-import { TRANSACTION_CATEGORY_TYPES } from '../constants/vet360';
+import { TRANSACTION_CATEGORY_TYPES } from '../../constants/vet360';
 import { selectVet360PendingCategoryTransactions } from '../selectors';
 
 function Vet360PendingTransactionCategory({ refreshTransaction: dispatchRefreshTransaction, transactions, hasPendingCategoryTransaction, categoryType, children }) {
