+++ src/applications/personalization/profile360/vet360/containers/TransactionReporter.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import scrollToTop from '../../../../platform/utilities/ui/scrollToTop';
+import scrollToTop from '../../../../../platform/utilities/ui/scrollToTop';
 
 import {
   selectVet360FailedTransactions,
@@ -11,9 +11,9 @@ import {
 
 import {
   clearTransaction
-} from '../actions';
+} from '../../actions';
 
-import Vet360TransactionErrorBanner from '../components/Vet360TransactionErrorBanner';
+import Vet360TransactionErrorBanner from '../components/base/TransactionErrorBanner';
 
 class Vet360TransactionReporter extends React.Component {
   static propTypes = {
