+++ src/applications/personalization/profile360/vet360/tests/components/TransactionPending.unit.spec.jsx
@@ -3,7 +3,7 @@ import enzyme from 'enzyme';
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import Vet360TransactionPending from '../../components/Vet360TransactionPending';
+import Vet360TransactionPending from '../../components/base/TransactionPending';
 
 describe('<Vet360TransactionPending/>', () => {
