+++ src/applications/personalization/profile360/vet360/tests/containers/CopyMailingAddress.unit.spec.jsx
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import {
   FIELD_NAMES
-} from '../../constants/vet360';
+} from '../../../constants/vet360';
 
 import {
   mapStateToProps
