+++ src/applications/personalization/profile360/vet360/tests/containers/ProfileField.unit.spec.jsx
@@ -3,8 +3,8 @@ import enzyme from 'enzyme';
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { Vet360ProfileField } from '../../containers/Vet360ProfileField';
-import { TRANSACTION_STATUS } from '../../constants/vet360';
+import { Vet360ProfileField } from '../../containers/ProfileField';
+import { TRANSACTION_STATUS } from '../../../constants/vet360';
 
 function Content() {
   return <h1>Content</h1>;
